Backend/python-backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Database dependencies
PyMySQL==1.1.0
SQLAlchemy==2.0.23
mysql-connector-python==9.3.0

# Authentication and security
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0

# Computer Vision and Biometrics
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
scikit-image==0.22.0

# Machine Learning and Analytics
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
joblib==1.3.2

# Data processing
python-dateutil==2.8.2
pytz==2023.3

# HTTP and API
requests==2.31.0
urllib3==2.1.0

# Logging and monitoring
python-json-logger==2.0.7

# File handling
python-magic==0.4.27

# Environment and configuration
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Health and safety specific
reportlab==4.0.7
openpyxl==3.1.2

# AI and NLP (optional)
openai==0.27.8
transformers==4.36.2
torch==2.1.2
